requirements.txt
Pillow>=9.1.1
tensorflow>=2.11.0